+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - New `args` argument in `core.odm.Model.t()`
 - New events in ODM model: `odm.entity.pre_save`, `odm.entity.save` 
 - New events in Content ODM model: `content.entity.pre_save`, `content.entity.save`
-- New properties in Content ODM model: `title`, `description`, `body`, `url`, `tags` 
+- New properties in Content ODM model: `title`, `description`, `body`, `url`, `tags`, `images`
 - New properties in Taxonomy ODM model: `title`, `alias`, `language`, `weight`, `order` 
 - New `stage` argument in `pytsite.odm_ui.UIMixin.setup_m_form()` hook.
 - New `args` argument in `pytsite.core.odm.Model.t()`.